fix: memisahkan example dari lib.rs
